EasyProg: Torture test extended with Flash Write and RAM

+++ EasyProg/include/flashcode.h
@@ -12,7 +12,7 @@
 
 void __fastcall__ flashCodeSetBank(uint8_t nBank);
 uint8_t __fastcall__ flashCodeGetBank(void);
-unsigned __fastcall__ flashCodeReadIds(uint8_t* pBase);
+uint16_t __fastcall__ flashCodeReadIds(uint8_t* pBase);
 void __fastcall__ flashCodeSectorErase(uint8_t* pBase);
 void __fastcall__ flashCodeWrite(uint8_t* pAddr, uint8_t nVal);
 uint8_t __fastcall__ flashCodeCheckProgress(uint8_t* pAddr);
